+++ internal/ingress/controller/store/store.go
@@ -328,7 +328,7 @@ func New(
 			return
 		}
 
-		if isCatchAllIngress(ing.Spec) && disableCatchAll {
+		if hasCatchAllIngressRule(ing.Spec) && disableCatchAll {
 			klog.InfoS("Ignoring delete for catch-all because of --disable-catch-all", "ingress", klog.KObj(ing))
 			return
 		}
@@ -353,7 +353,7 @@ func New(
 				return
 			}
 
-			if isCatchAllIngress(ing.Spec) && disableCatchAll {
+			if hasCatchAllIngressRule(ing.Spec) && disableCatchAll {
 				klog.InfoS("Ignoring add for catch-all ingress because of --disable-catch-all", "ingress", klog.KObj(ing))
 				return
 			}
@@ -377,7 +377,7 @@ func New(
 			validOld := class.IsValid(oldIng)
 			validCur := class.IsValid(curIng)
 			if !validOld && validCur {
-				if isCatchAllIngress(curIng.Spec) && disableCatchAll {
+				if hasCatchAllIngressRule(curIng.Spec) && disableCatchAll {
 					klog.InfoS("ignoring update for catch-all ingress because of --disable-catch-all", "ingress", klog.KObj(curIng))
 					return
 				}
@@ -389,7 +389,7 @@ func New(
 				ingDeleteHandler(old)
 				return
 			} else if validCur && !reflect.DeepEqual(old, cur) {
-				if isCatchAllIngress(curIng.Spec) && disableCatchAll {
+				if hasCatchAllIngressRule(curIng.Spec) && disableCatchAll {
 					klog.InfoS("ignoring update for catch-all ingress and delete old one because of --disable-catch-all", "ingress", klog.KObj(curIng))
 					ingDeleteHandler(old)
 					return
@@ -624,10 +624,10 @@ func New(
 	return store
 }
 
-// isCatchAllIngress returns whether or not an ingress produces a
+// hasCatchAllIngressRule returns whether or not an ingress produces a
 // catch-all server, and so should be ignored when --disable-catch-all is set
-func isCatchAllIngress(spec networkingv1beta1.IngressSpec) bool {
-	return spec.Backend != nil && len(spec.Rules) == 0
+func hasCatchAllIngressRule(spec networkingv1beta1.IngressSpec) bool {
+	return spec.Backend != nil
 }
 
 // syncIngress parses ingress annotations converting the value of the